test: split test creators from actual test code
